feat!: bump @serlo/editor & make editor create initial content & refactor
