+++ docker-compose.yml
@@ -19,12 +19,6 @@ services:
       - 3000:3000
     develop:
       watch:
-        # TODO: Sync would be faster but harder to setup.
-        # - action: sync
-        #   path: ./src
-        #   target: /usr/src/app/src
-        #   ignore:
-        #     - node_modules/
         - action: rebuild
           path: .
     extra_hosts:
@@ -48,11 +42,6 @@ services:
       interval: 10s
       timeout: 5s
       retries: 3
-    # image: ghcr.io/serlo/serlo-mysql-database:latest
-    # platform: linux/x86_64
-    # pull_policy: always
-    # ports:
-    #   - '3306:3306'
     networks:
       - express-mongo-mysql
 networks:
