+++ package.json
@@ -28,7 +28,7 @@
   },
   "dependencies": {
     "@dotenvx/dotenvx": "^1.14.2",
-    "@serlo/editor": "0.15.4",
+    "@serlo/editor": "0.16.0",
     "express": "^4.19.2",
     "fp-ts": "2.16.9",
     "io-ts": "^2.2.21",
