+++ src/frontend/App.tsx
@@ -118,6 +118,7 @@ function App() {
           })
       })
     }
+    // eslint-disable-next-line react-hooks/exhaustive-deps
   }, [])
 
   if (appState.type === 'fetching-content') return null
